refactor(codeowners): make sync/check default trait methods
The GitHub/GitLab/Bitbucket providers carried byte-identical sync()
and check() bodies differing only in output_path() and
section_style(). Lift those bodies into default methods on
CodeOwnersProvider and share a normalize_content helper with
write_codeowners_file; each platform impl is now just its two
accessors.

Validation: cargo test -p cuenv-codeowners/-gitlab/-bitbucket/-github
(153 tests green), all-target clippy clean, cargo fmt, git diff
--check.
